pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "microplas-blind-corr"
version = "2.0.0"
description = "General-purpose tool for microplastics particle data processing with blank and blind correction"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Robert Naumann"},
]
keywords = ["microplastics", "environmental", "data-processing", "quality-control"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.2",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pyyaml>=5.4.0",
    "jupyter>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.900",
]
jupyter = [
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/robna/gepard_BlindCorr"
Repository = "https://github.com/robna/gepard_BlindCorr.git"
Issues = "https://github.com/robna/gepard_BlindCorr/issues"

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "configs/",
    "examples/",
    "tests/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/microplas_blind_corr"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src/microplas_blind_corr"
testpaths = [
    "tests",
]
